requirements.txt
python-docx==0.8
pyttsx3==2.90